+++ docs/source/whatsnew.rst
@@ -29,8 +29,11 @@ Legend
 
 Version 0.1.0
 ==============
+- |Remove| the embedding model Universal Sentence Encoder (USE) and its
+  dependencies (:code:`tensorflow` and :code:`tensorflow-hub`).
 - |Remove| :code:`BBS_BBG_poc` notebook (now hosted on https://github.com/BlueBrain/Search-Graph-Examples ) and
 :code:`assets/` directory
+
 
 Version 0.0.10
 ==============